Optimise Order Profiles Card and Modal View components.
