+++ packages/legend-application-marketplace/src/components/ProviderCard/orderProfileUtils.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2025-present, Goldman Sachs
+ * Copyright (c) 2026-present, Goldman Sachs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,11 +15,7 @@
  */
 
 import type { TraderProfileItem } from '@finos/legend-server-marketplace';
-
-// ─── Category constants ───────────────────────────────────────────────────────
-
-/** Canonical lower-case value stored in TraderProfileItem.category for terminals. */
-export const VENDOR_PROFILE_CATEGORY = 'vendor profile';
+import { MAX_PRODUCT_IMAGE_COUNT } from '../../stores/lakehouse/dataProducts/ProductCardState.js';
 
 // ─── String labels ───────────────────────────────────────────────────────────
 
@@ -37,6 +33,7 @@ export enum OrderProfileLabel {
   IN_CART_SUFFIX = '(In Cart)',
   MODEL_PREFIX = 'Model: ',
   PRICE_TOTAL_SEPARATOR = ' · Total: ',
+  VIEW_DETAILS = 'View details',
 }
 
 export enum OrderProfileTableHeader {
@@ -45,6 +42,18 @@ export enum OrderProfileTableHeader {
   CATEGORY = 'CATEGORY',
   COST_MONTHLY = 'COST (Monthly)',
 }
+
+// ─── Image URL ──────────────────────────────────────────────────────────────────
+
+/**
+ * Returns a stable random product image URL for a given asset base URL.
+ * Intended to be called once inside a `useState` initialiser so the image
+ * does not change on re-renders.
+ */
+export const getRandomImageUrl = (assetUrl: string): string => {
+  const randomIndex = Math.floor(Math.random() * MAX_PRODUCT_IMAGE_COUNT) + 1;
+  return `${assetUrl}/images${randomIndex}.jpg`;
+};
 
 // ─── Price formatting ─────────────────────────────────────────────────────────
 
@@ -63,6 +72,16 @@ export const formatItemPrice = (price: number): string =>
 export const formatCardPrice = (price: number): string =>
   `${USD_FORMATTER.format(price)}/month`;
 
+// ─── Toast message formatters ────────────────────────────────────────────────
+
+export const formatAddToCartSuccessMessage = (productName: string): string =>
+  `Order profile ${productName} has been successfully added to cart.`;
+
+export const formatAddToCartErrorMessage = (
+  productName: string,
+  errorMessage: string,
+): string => `Failed to add ${productName} to cart: ${errorMessage}`;
+
 // ─── Item summary helpers ─────────────────────────────────────────────────────
 
 export const getItemSummary = (
@@ -80,6 +99,30 @@ export const formatProfileSummaryLine = (
     terminalCount === 1 ? '1 Terminal' : `${terminalCount} Terminals`;
   const addOnLabel = addOnCount === 1 ? '1 Add-On' : `${addOnCount} Add-Ons`;
   return `${terminalLabel} · ${addOnLabel}`;
+};
+
+// ─── Multiselect price calculation ──────────────────────────────────────────
+
+/**
+ * Calculates the total price for a multiselect order profile.
+ * Finds the highest-priced terminal and sums it with its associated add-ons.
+ * Returns `undefined` when there are no terminal items.
+ */
+export const calculateMultiselectTotalPrice = (
+  items: TraderProfileItem[],
+): number | undefined => {
+  const terminals = items.filter((item) => item.isTerminal);
+  if (terminals.length === 0) {
+    return undefined;
+  }
+  const highestTerminal = terminals.reduce((max, curr) =>
+    curr.price > max.price ? curr : max,
+  );
+  const addOns = items.filter(
+    (item) => !item.isTerminal && item.model === highestTerminal.model,
+  );
+  const addOnsTotal = addOns.reduce((sum, item) => sum + item.price, 0);
+  return highestTerminal.price + addOnsTotal;
 };
 
 // ─── Grouping ─────────────────────────────────────────────────────────────────
